Use .key instead of deprecated .keyCode, fix #10
Apparently [.keyCode is deprecated](https://stackoverflow.com/a/45771014) and there is also a bug where users can't type the letter q when using the extension. This commit replaces `.keyCode` with the use of `.key`, which should fix the bug and still keep both variants of PwdHash (prefix @@ or password key F2) working. Needs testing.